Now passes!!

+++ xtrack/beam_elements/elements_src/track_magnet.h
@@ -210,8 +210,8 @@ void track_magnet_body_single_particle(
             if (num_multipole_kicks > 1) {
                 edge_drift_weight = 1. / (2 * (1 + num_multipole_kicks));
                 inside_drift_weight = (
-                    ((float) num_multipole_kicks)
-                        / ((float)(num_multipole_kicks*num_multipole_kicks) - 1));
+                    ((double) num_multipole_kicks)
+                        / ((double)(num_multipole_kicks*num_multipole_kicks) - 1));
             }
 
             MAGNET_DRIFT(part, edge_drift_weight*length);
